+++ now/constants.py
@@ -7,7 +7,7 @@ from now.utils import BetterEnum
 # ----------------------------------
 DOCKER_BFF_PLAYGROUND_TAG = '0.0.138-fix-search-options-order1'
 # ----------------------------------
-NOW_PREPROCESSOR_VERSION = '0.0.117-test-local-end-to-end-15'
+NOW_PREPROCESSOR_VERSION = '0.0.117-test-local-end-to-end-16'
 NOW_QDRANT_INDEXER_VERSION = '0.0.138-refactor-remove-finetuning-6'
 NOW_ELASTIC_INDEXER_VERSION = '0.0.139-test-local-end-to-end-12'
 NOW_AUTOCOMPLETE_VERSION = '0.0.7-refactor-remove-finetuning-6'
